(git.install) returned code that adds to PATH

+++ automatic/git.install/tools/chocolateyInstall.ps1
@@ -22,4 +22,13 @@ $packageArgs = @{
     ValidExitCodes = @(0)
 }
 Install-ChocolateyInstallPackage @packageArgs
+
+$packageName = $packageArgs.packageName
+$installLocation = Get-AppInstallLocation $packageArgs.SoftwareName
+if (!$installLocation)  { Write-Warning "Can't find $packageName install location"; return }
+Write-Host "$packageName installed to '$installLocation'"
+
+Install-ChocolateyPath "$installLocation\cmd"
+
+
 rm -Force $filePath32, $filePath64 -ea 0
